reading_section/requirements.txt
SQLAlchemy==2.0.41
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
passlib==1.7.4
pydantic==2.11.7
pydantic_core==2.35.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
openai
passlib[bcrypt]
python-jose[cryptography]